Improve ergonomics of using TheRock as a development driver for sub-projects. (#1242)
The motivation of this change is to enable the following developer
productivity features:

* Development in a sub-project `build` directory should cause the
super-project to see the sub-project as out of date and take appropriate
action. This is done by injecting an ALL `therock-touch` target into
each sub-project so that `ninja` will invalidate the super-project
stage-install state, causing a re-install and rebuild of dependents.
* One stop commands to build and distribute in a sub-project without
switching back and forth. Users can now just run `ninja therock-dist` in
any sub-project to build and distribute their local changes (i.e. to
`build/dist/rocm`, etc).
* More coverage of `expunge` targets, making sure that project state can
be reset more completely.

Key changes:

* Removes some `therock-` prefixes from top level convenience targets.
* Defines `artifacts`, `archives`, `dist`, and `expunge` top level
convenience targets explicitly.
* Adds `expunge` steps to all phases of build, artifact assembly and
dist population.
* Removes the formal sub-project `+dist` phase, collapsing the creation
of the local `dist` directory into the `+stage` (install) step. This
separation was done early on when creating the dist/ tree was slow but
it doesn't pay for itself anymore and was becoming hard to distinguish
the different "dist" concepts in the project.
* Artifacts can be declared as belonging to different distributions
(default "rocm"), resulting in population of different trees in
`build/dist/{distribution}`. This was a TODO from the beginning.
* `{subproject}+dist` now is a convenience target that builds all
artifacts the project is a part of and populates the
`build/dist/{distname}` directory.
* Distributions (i.e. `build/dist/rocm`) are now populated incrementally
as artifact dependencies are met instead of as part of one top level
target which requires the whole project to be (re)built.
* New targets `therock-touch` and `therock-dist` are injected into every
sub-project, allowing development to be done completely in a sub-project
build directory while keeping super-project level artifacts and
distributions populated.

This should be NFC for all automation since the new barename targets
(dist, archives) are aliased to their originals. We could simplify this
in followups.