[ci] - Adding fix for artifact pull (#1270)
Noticed failures in TheRock CI. Due to GitHub's strange typing in
workflows, the script always considered pulled artifacts to be external.
(boolean for GitHub Actions = string in Python)

This fix pulls the correct artifacts. I decided to do TheRock fix
instead of forked fix, since this is easier to test on workflows and
prove it is working, since this is breaking the current workflow

working here: https://github.com/ROCm/TheRock/actions/runs/17050492399