Add documentation for torch wheel versions (#1385)
## Motivation

By default, the nightly wheel install command we provide currently pulls
in incompatible wheels. To address this we need to communicate which
wheels are compatible, which serves the dual purpose of communicating
that we provide nightly builds for torch versions other than the default
versions pip pulls in.

Currently this change adds --pre to pull the latest prerelease wheels by
default and provides one alternate set of compatible wheels for torch 2.7.1.
In the future we may want a proper compatibility matrix so we can scale
properly as future torch versions are released.

## Test Plan

Specifying --pre or torchaudio==2.7.1a0 should resolve the bad_alloc issue seen in
https://github.com/ROCm/TheRock/issues/1378.

## Test Result

Both provided sets of compatible wheels resolve the issue.