[torch] Flip `--enable-pytorch-flash-attention-windows` for releases. (#1437)
## Motivation

Fixes https://github.com/ROCm/TheRock/issues/1040, enabling aotriton for
flash attention in pytorch (if it works). This is expected to improve
performance in workloads like ComfyUI image generation by upwards of 60%
(e.g. 12.6 it/s to 20.0 it/s).

## Technical Details

Follow-up to https://github.com/ROCm/TheRock/pull/1432 and depends on
https://github.com/pytorch/pytorch/pull/162330.

Note that support is experimental for some GPUs like gfx1100, so the
`TORCH_ROCM_AOTRITON_ENABLE_EXPERIMENTAL=1` environment variable may be
needed to try aotriton on those systems.

## Test Plan

Trigger either
https://github.com/ROCm/TheRock/actions/workflows/build_windows_pytorch_wheels.yml
or
https://github.com/ROCm/TheRock/actions/workflows/release_windows_pytorch_wheels.yml
across the matrix of GPU families once that PyTorch PR is merged.

We're still going to need automated tests and documentation for this.
I'd like numerics tests running somewhere and documentation that shows
how to check which pytorch features are enabled in the wheels that a
user installs.

## Test Result

Test runs:

* https://github.com/ROCm/TheRock/actions/runs/17660396787 using this
branch and `7.0.0rc20250908` for gfx110X-dgpu
* ~~https://github.com/ROCm/TheRock/actions/runs/17660456285 using the
branch and `7.0.0rc20250908` for gfx1151~~
* https://github.com/ROCm/TheRock/actions/runs/17662170140 using the
branch and `7.0.0rc20250908` for gfx1151
* Tests not running should be fixed with
https://github.com/ROCm/TheRock/pull/1469

(may need to retrigger to pick up fixes for flaky checkouts)

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.