Bump CMake version to 3.27 in manylinux image to match PyTorch's (#1512)
Bumping the CMake version of the manylinux image used by the CI to match
PyTorch's minimum CMake version 3.27. This is required since commit:
https://github.com/pytorch/pytorch/commit/c2beeadeb40e927c51bc5bcd409b3f28374a5190

The change is a requirement when building PyTorch nightly wheels.

This should be considered to be the new default policy for our CMake
progression for the manylinux image to match the PyTorch CMake
progression (if it is higher than ours).