Refactor artifact builder to make it more strict. (#1483)
We've been getting reports of some files missing from distribution
artifacts (notably for rocprofiler, which added some additional python
files recently).

Since the tool that does artifact building was the first thing written 9
months ago, and since it was a bit of a pragmatic attempt, I took the
chance to rewrite it so that I could enforce stricter invariants and
catch more issues at build time.

Functional changes:

* Components now "extend" another component and a file can only match
the deepest component in its hierarchy. Only unmatched files are
eligible for child components. This replaces the carefully balanced
include/exclude defaults which attempted to ensure that each component's
patterns represented a mutually exclusive set. Now mutual exclusion is
by construction.
* Prior, the build scripts would run the artifact builder once per
component, but now all components in a descriptor are processed in one
invocation. This allows us to form a global view of what goes where once
and then do the populating all in one go.
* The above enables stricter verification options. Particularly,
unmatched files which are not included in any component are now tracked.
By default, any unmatched component will produce a build error. The
error message contains the changes needed to declare them. All files
must either be matched by a component or declared in the descriptor
options to be ok to ignore.
* The tighter checks have revealed a number of descriptors which were
leaving some of their contents out. These have been fixed (and more that
the CI surfaces will be fixed prior to landing).

Other changes:

* Complete unit tests have been implemented for the artifact builder.
This was being tested only by a couple of very course integration tests.
* The artifact cmake scripts was inadvertently invoking the flatten tool
multiple times as directories were accumulated, with each invocation
adding one more directory to the list of things to flatten. This was due
to improper nesting in a foreach and has been reduced to one invocation.
* Strict verification of all descriptor fields and tables was added.

Fixes https://github.com/ROCm/rocm-libraries/issues/1622