Document s3_management folder and iterate on script behavior. (#1530)
We forked these scripts from
https://github.com/pytorch/test-infra/tree/main/s3_management where they
were not really documented. We have them partially running via
automation but some tasks still need the scripts to be run manually.
I've documented the current state. As for automating more of this, see
also the upstream issues:
https://github.com/pytorch/test-infra/issues/6097 and
https://github.com/pytorch/test-infra/issues/6105.

Tested by following the steps to upload `setuptools` to
https://rocm.nightlies.amd.com/v2-staging/gfx1151, which should fix this
failure:
https://github.com/ROCm/TheRock/actions/runs/17819440813/job/50662249490
```
  Looking in indexes: https://rocm.nightlies.amd.com/v2-staging/gfx1151
  ERROR: Could not find a version that satisfies the requirement setuptools>=70.2.0 (from versions: none)
  ERROR: No matching distribution found for setuptools>=70.2.0
```

---------

Co-authored-by: Marius Brehler <marius.brehler@gmail.com>