[flagflip] Flip THEROCK_FLAG_COMGR_DELAY_LOAD=ON. (#1569)
* NFC for Windows.
* After many hours of debugging, I've concluded that there is a latent
heap corruption in the RTC compiler path used by rocfft. This
semi-consistently crashes on RHEL8 at build time with a segfault and
unrecoverable stack trace after dereferencing a nullptr. Something about
the timing or allocation ordering causes this to recur in a namespace vs
not. After a very lengthy (and ultimately inconclusive) root cause
debugging session, this is believed to be a latent bug vs introduced by
these changes.
* Given the above, I have amended this patch with more hardening of the
implementation and introduced the new namespacing behavior as an opt-in
environment variable that will be activated when AMD_COMGR_NAMESPACE=1.
We can make this the default once a full root cause/fix is established,
but landing it now gives an option for users who need additional safety
with respect to how LLVM is used.
* Having the namespace code path compiled in but disabled will aid in
further root causing efforts.

Progress on #1543