[CI:] Use python script for health status (#1516)
New python script that will replace the current health status in the CI.
It is an extension of build_tools/hack/diagnose.py that unifies all
status information of a compute node relevant for TheRock. This includes
system information like cpu and storage, ccache status and config,
required software and their versions (gcc, python,..), and the list of
installed python packages.

Related issue #1456