Handle promotion to v2 from v2-staging by disabling upload_pytorch_wheels step run when workflow is cancelled (#1579)
An issue was seen with the upload_pytorch_wheels gating check added
through PR #1072 and #1382 .

Promotion to v2 from v2-staging should be blocked when workflow is
cancelled. Instead through the promote_wheels_based_on_policy.py helped
script we end up uploading the wheels even if it is not tested to v2
from v2 staging if upload flag is determined to be true here.


https://github.com/ROCm/TheRock/blob/82d23dd64ad23e7ee9915240823375641661125d/build_tools/github_actions/promote_wheels_based_on_policy.py#L19-L44
falling through to the # 4) Otherwise → upload=true

This PR makes the changes in the workflow to handle cancelled workflow
for both linux and windows.

---------

Co-authored-by: arravikum <arravikum@amd.com>