SWDEV-555589 - Edit libelf, libdw and libasm source code to create libraries with updated name (#1576)
The renamed and patchelf edited libraries in ROCk artifacts are not
providing the symbol version tag. This was resulting in rpm package
installation failure.
Patch will create the library with the updated name. Patchelf editing is
not required.
Create libelf, libdw and libasm libraries with "librocm_sysdeps_" as
prefix.

## Motivation

The patchelf edited libraries are not providing the symbol version tag

## Technical Details

Rather than editing libraries to update soname , create libraries with
the updated name by editing the target name in source code

## Test Plan

Verified the elfdeps provides field of the libraries.

## Test Result

Required symbol version tags are provided by the libraries

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Co-authored-by: raramakr <raramakr@amd.com>