[ci] Disable mi355 test machines (#1661)
Currently, mi355 tests are incredibly flaky due to networking issues
(which is a cluster issue). Worked with OSSCI to figure out the root
cause, however, it seems to be a conductor mi355 issue that is causing
issues with other teams.

This issue may not be resolved for some time, so to prevent flaky
results, we will be disabling for the time being

Issue to reenable #1660