Add progress printing for git submodule updates (#1574)
## Motivation

By default git submodule does not display download progress. This small
feature allows users to monitor downloads to insure downloads have not
stalled. This is a benefit for the larger submodules that can take some
time to download.

## Technical Details

This patch adds an option, --progress, to the script arguments. When
passed to the python script it is added to arguments passed to the
submodule command.

## Test Plan

Tested running fetch sources with the --progress flag passed.

## Test Result

Flag works as intended.