Add fftw3 and fftw3f 3.3.10 libraries (#1573)
Add the latest 3.3.10 fftw3 (double precision) and fftw3f (float/single
precision) libraries to TheRock.
Libraries are build with the multi-thread support.

Libraries are build with the cmake and in this way they will
automatically include the FFTW3Configure.cmake and FFTW3fConfigure.cmake
and other related cmake files that can be used by cmake to find these libraries.

When libraries are found, these scripts set variables like
FFTW3_FOUND and FFTW3f_FOUND and target properties FFTW3::fftw3 and
FFTW3::fftw3f.

These cmake helper files are not produced if the fftw is build with the 
autoconf/configure/make way which is the reason why they are not
included in the fftw-packages that many linux distributions provide.
In those cases applications often use custom FindFFTW.cmake where the
variables set are different. (For exampe FFTW_FOUND instead of the FFTW3_FOUND,
FFTW_LIBRARY_DIRS, etc...)

---------

Signed-off-by: Mika Laitio <mika.laitio@amd.com>
Co-authored-by: Geo Min <geomin12@amd.com>