[RFC] RFC0004 proposing adding Fusilli+IREE as kernel provider and JIT engine for hipDNN (#1821)
This RFC proposes adding IREE as a kernel provider to hipDNN to leverage
JIT compiled and codegenerated kernels in ML training and inference
solutions. This is made possible with the development of Fusilli - a C++
graph API and JIT engine for IREE. We believe hand-authored kernel
libraries are great for highly tuned performance but they are difficult
to 1) scale to newer models or target architectures and 2) package and
release effectively. This RFC is founded on the overarching goal to
complement our software stack with JIT solutions while being competitive
to hand-authored kernel libraries. Apart from the usual benefits of
having a compiler-backed JIT engine that gets progressively better, a
systemic benefit of this is it helps reduce build times and binary
sizes, making it easier to ship software effectively.