[Documentation] Update python_packaging.md (#1841)
## Motivation

The example is missing the version flag letting users copy and paste the
commands to build. This flag has a default that causes `pip install` to
fail down the line (saying it can't find a valid version) that requires
the user to dig into the scripts to figure out why.