[ci] Skipping jobs based on failure and fixing CI Nightly build variants (#1898)
After #1766 , CI works great. However, CI Nightly was never updated,
resulting in no jobs:
https://github.com/ROCm/TheRock/actions/runs/18735688316

In another issue, ASAN builds typically fail but is marked as xfail. In
that case, it still considers it as a "success" and carries on with
future jobs. For this workflow:
https://github.com/ROCm/TheRock/actions/runs/18736946879/job/53446872470,
that is not the case. This PR patches that too