[ci] Fixing failing tests and fixing `CI summary` (#1942)
After rocm-libraries merge, there are some new test timeouts:
https://github.com/ROCm/TheRock/actions/runs/18794083478

For `hipblas` and `rocsparse`, we are extending the timeout. For MIOpen,
we correctly ignore the test for smoke-tests, but not for full-tests -
this PR fixes that

For `hipblaslt`, we see a few instances of timeouts in rocm-libraries,
so we extend the timeout to ensure those tests get enough time:
-
https://github.com/ROCm/rocm-libraries/actions/runs/18793151505/job/53632054964
-
https://github.com/ROCm/rocm-libraries/actions/runs/18784309594/job/53606333527

Also for `CI Summary`, it doesn't include `continue-on-error` jobs thus
marking it as failure. Now, this will correctly ignore it and make it
green