Python script to retrieve PRs changing the same files (#1917)
Returns list of PRs that change the same files matching a given pattern.
Pattern is applied as partial matching for the full path name (=
directory name + file name).

Repo owner and repo name are modifiable, so it can also be used to
quickly check if e.g. in rocm-libraries PRs are touching the same files
you are working on.

GitHub fine-grained personal access token is required to make it work.
Rate limits from GitHub will limit how often this can be run in short
time.